+++ CHANGELOG.md
@@ -6,6 +6,11 @@ For more information about changelogs, check
 [Keep a Changelog](http://keepachangelog.com) and
 [Vandamme](http://tech-angels.github.io/vandamme).
 
+## 0.1.3 - 1/31/17
+
+* [ENHANCEMENT] YtPotion.Video and YtPotion.Channel modules to return the HTTPoison response
+* [UPDATE] Update YtPotion.Parser method name(s)
+
 ## 0.1.2 - 1/27/17
 
 * Update YtPotion.Video and YtPotion.Channel modules to return lists of map data instead of a HTTPoison response